AIFitnessCoachApp/backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.7

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0
greenlet==3.0.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# AI & Agent Framework
openai==1.58.1
openai-agents-sdk==0.1.1
anthropic==0.39.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.3.0
crewai==0.70.1
agentsdk==0.1.0

# MCP Integration
mcp==0.1.0
anthropic-mcp-sdk==0.1.0

# Advanced AI Tools
tiktoken==0.5.2
chromadb==0.4.22
sentence-transformers==2.2.2

# HTTP Client
httpx==0.26.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Caching
redis==5.0.1
aiocache==0.12.2

# Task Queue
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.5
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Exercise Database
requests==2.31.0

# JSON Logging
python-json-logger==2.0.7